[FE] REFACTOR: 디렉토리 comment 분기 및 불필요한 코드 삭제
